+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add feature to open file or folder from command arguments
 - Add feature to open a specific file from the command palette
+- Add feature to create a new file or directory
 
 ## [0.0.2] - 2025-01-07